rename: 컴포넌트 폴더 케밥케이스로 통일
