+++ src/components/step-progress/StepProgress.jsx
