+++ src/components/step-progress/StepProgress.module.css
