+++ src/features/analysis-history/pages/AnalysisHistory.jsx
@@ -2,8 +2,8 @@ import { useRef } from 'react'
 import cx from 'classnames'
 import { useParams } from 'react-router-dom'
 
-import { BackButton } from '@/components/BackButton/BackButton'
-import { ScrollToTopButton } from '@/components/ScrollToTopButton/ScrollToTopButton'
+import { BackButton } from '@/components/back-button/BackButton'
+import { ScrollToTopButton } from '@/components/scroll-to-top-button/ScrollToTopButton'
 import { ContractAnalysisLoading } from '@/features/contract-analysis-result/components/ContractAnalysisLoading'
 import { useScrollSnap } from '@/hooks/useScrollSnap'
 import { useStep } from '@/stores/useStep'
